First window tests working, see below for details
First step to fully replace compute shader with Zig code running on the CPU.

This was started because of misunderstandings wrt. storage textures (fml)
and shortcomings of the Mach shader compiler that I'm not able to easily fix myself.
The shader was poorly written and Mach plans to replace WGSL with Zig anyways,
so a rewrite was needed sooner or later anyways.